feat(enhancement): Conversation apply nodes are now able to grant payments, trigger events, write log entries, and more (#5157)
An `apply` node can now do anything that a mission's `on *` blocks can do except create dialogs or conversation or require outfits.
