+++ source/Conversation.cpp
@@ -86,7 +86,7 @@ bool Conversation::RequiresLaunch(int outcome)
 
 
 // Load a conversation from file.
-void Conversation::Load(const DataNode &node)
+void Conversation::Load(const DataNode &node, const string &missionName)
 {
 	// Make sure this really is a conversation specification.
 	if(node.Token(0) != "conversation")
@@ -171,7 +171,7 @@ void Conversation::Load(const DataNode &node)
 			// Don't merge "apply" nodes with any other nodes.
 			AddNode();
 			nodes.back().canMergeOnto = false;
-			nodes.back().conditions.Load(child);
+			nodes.back().actions.Load(child, missionName);
 		}
 		// Check for common errors such as indenting a goto incorrectly:
 		else if(child.Size() > 1)
@@ -241,17 +241,22 @@ void Conversation::Save(DataWriter &out) const
 				out.Write("scene", node.scene->Name());	
 			if(!node.conditions.IsEmpty())
 			{
-				// The only thing differentiating a "branch" from an "apply" node
-				// is that a branch has two data entries instead of one.
-				if(node.data.size() > 1)
-					out.Write("branch", TokenName(node.data[0].second), TokenName(node.data[1].second));
-				else
-					out.Write("apply", TokenName(node.data[0].second));
-				
+				out.Write("branch", TokenName(node.data[0].second), TokenName(node.data[1].second));
 				// Write the condition set as a child of this node.
 				out.BeginChild();
 				{
 					node.conditions.Save(out);
+				}
+				out.EndChild();
+				continue;
+			}
+			if(!node.actions.IsEmpty())
+			{
+				out.Write("apply", TokenName(node.data[0].second));
+				// Write the GameAction as a child of this node.
+				out.BeginChild();
+				{
+					node.actions.SaveAction(out);
 				}
 				out.EndChild();
 				continue;
@@ -301,13 +306,36 @@ bool Conversation::IsValidIntro() const noexcept
 
 
 
-// Do text replacement throughout this conversation.
-Conversation Conversation::Substitute(const map<string, string> &subs) const
+// Check if the actions in this conversation are valid.
+string Conversation::Validate() const
+{
+	for(const Node &node : nodes)
+	{
+		if(!node.actions.IsEmpty())
+		{
+			string reason = node.actions.ValidateAction();
+			if(!reason.empty())
+				return "conversation applied " + reason;
+		}
+	}
+	return "";
+}
+
+
+
+// Do text replacement throughout this conversation, and instantiate any
+// potential actions.
+Conversation Conversation::Instantiate(map<string, string> &subs, int jumps, int payload) const
 {
 	Conversation result = *this;
 	for(Node &node : result.nodes)
+	{
 		for(pair<string, int> &choice : node.data)
 			choice.first = Format::Replace(choice.first, subs);
+		if(!node.actions.IsEmpty())
+			node.actions = node.actions.Instantiate(subs, jumps, payload);
+	}
+
 	return result;
 }
 
@@ -335,37 +363,47 @@ int Conversation::Choices(int node) const
 
 
 
-// Check if the given converation node is a conditional branch.
+// Check if the given conversation node is a conditional branch.
 bool Conversation::IsBranch(int node) const
 {
 	if(static_cast<unsigned>(node) >= nodes.size())
 		return false;
 	
-	return !nodes[node].conditions.IsEmpty() && nodes[node].data.size() > 1;
+	return !nodes[node].conditions.IsEmpty();
 }
 
 
 
-
-// Check if the given converation node applies changes to condition variables.
+// Check if the given conversation node applies an action.
 bool Conversation::IsApply(int node) const
 {
 	if(static_cast<unsigned>(node) >= nodes.size())
 		return false;
 	
-	return !nodes[node].conditions.IsEmpty() && nodes[node].data.size() == 1;
+	return !nodes[node].actions.IsEmpty();
 }
 
 
 
-// Get the list of conditions that the given node tests or applies.
-const ConditionSet &Conversation::Conditions(int node) const
+// Get the list of conditions that the given node tests for branching.
+const ConditionSet &Conversation::Branch(int node) const
 {
 	static ConditionSet empty;
 	if(static_cast<unsigned>(node) >= nodes.size())
 		return empty;
 	
 	return nodes[node].conditions;
+}
+
+
+// Get the action that the given node applies.
+const GameAction &Conversation::Apply(int node) const
+{
+	static GameAction empty;
+	if(static_cast<unsigned>(node) >= nodes.size())
+		return empty;
+	
+	return nodes[node].actions;
 }
