+++ source/ConversationPanel.cpp
@@ -331,16 +331,14 @@ void ConversationPanel::Goto(int index, int selectedChoice)
 		{
 			// Branch nodes change the flow of the conversation based on the
 			// player's condition variables rather than player input.
-			choice = !conversation.Conditions(node).Test(player.Conditions());
+			choice = !conversation.Branch(node).Test(player.Conditions());
 		}
 		else if(conversation.IsApply(node))
 		{
-			// Apply nodes alter the player's condition variables but do not
-			// display any conversation text of their own.
-			player.SetReputationConditions();
-			conversation.Conditions(node).Apply(player.Conditions());
-			// Update any altered government reputations.
-			player.CheckReputationConditions();
+			// Apply nodes are able to take various actions, be that changing
+			// the player's conditions, granting payments, triggering events,
+			// and more.
+			conversation.Apply(node).DoAction(player);
 		}
 		else
 		{
