+++ source/StartConditions.cpp
@@ -148,7 +148,7 @@ void StartConditions::FinishLoading()
 	for(Ship &ship : ships)
 		ship.FinishLoading(true);
 	
-	if(!GetConversation().IsValidIntro())
+	if(!GetConversation().IsValidIntro() || !GetConversation().Validate().empty())
 		Files::LogError("Warning: The start scenario \"" + Identifier() + "\" (named \""
 			+ GetDisplayName() + "\") has an invalid starting conversation.");
 }
@@ -166,7 +166,7 @@ bool StartConditions::IsValid() const
 		return false;
 	
 	// A start must reference a valid "intro" conversation, either stock or custom.
-	if((stockConversation && !stockConversation->IsValidIntro()) || (!stockConversation && !conversation.IsValidIntro()))
+	if(!GetConversation().IsValidIntro() || !GetConversation().Validate().empty())
 		return false;
 	
 	// All ship models must be valid.
